move package 'filter' from org.mybatis.spring.scan to org.mybatis.spring
