+++ src/test/java/org/mybatis/spring/filter/ScanFilterTest.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.mybatis.spring.scan.filter;
+package org.mybatis.spring.filter;
 
 import com.mockrunner.mock.jdbc.MockDataSource;
 import org.junit.jupiter.api.Test;
 import org.mybatis.spring.SqlSessionFactoryBean;
-import org.mybatis.spring.scan.filter.config.AppConfig;
+import org.mybatis.spring.filter.config.AppConfig;
 import org.springframework.beans.factory.support.GenericBeanDefinition;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
