+++ src/test/java/org/mybatis/spring/filter/config/AppConfig.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.mybatis.spring.scan.filter.config;
+package org.mybatis.spring.filter.config;
 
 import org.mybatis.spring.annotation.MapperScan;
-import org.mybatis.spring.scan.filter.customfilter.AnnoTypeFilter;
-import org.mybatis.spring.scan.filter.customfilter.ExcludeMaker;
-import org.mybatis.spring.scan.filter.customfilter.CustomTypeFilter;
+import org.mybatis.spring.filter.customfilter.ExcludeMaker;
+import org.mybatis.spring.filter.customfilter.AnnoTypeFilter;
+import org.mybatis.spring.filter.customfilter.CustomTypeFilter;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.FilterType;
 
@@ -26,32 +26,32 @@ import org.springframework.context.annotation.FilterType;
 public class AppConfig {
 
 
-  @MapperScan(basePackages = "org.mybatis.spring.scan.filter.datasource",
+  @MapperScan(basePackages = "org.mybatis.spring.filter.datasource",
     excludeFilters = {@ComponentScan.Filter(type = FilterType.CUSTOM, classes = CustomTypeFilter.class)})
   public static class CustomFilterConfig {
 
   }
 
-  @MapperScan(basePackages = "org.mybatis.spring.scan.filter.datasource",
+  @MapperScan(basePackages = "org.mybatis.spring.filter.datasource",
     excludeFilters = {@ComponentScan.Filter(type = FilterType.ANNOTATION, classes = AnnoTypeFilter.class)})
   public static class AnnoFilterConfig {
 
   }
 
-  @MapperScan(basePackages = "org.mybatis.spring.scan.filter.datasource",
+  @MapperScan(basePackages = "org.mybatis.spring.filter.datasource",
     excludeFilters = {@ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = ExcludeMaker.class)})
   public static class AssignableFilterConfig {
 
   }
 
-  @MapperScan(basePackages = "org.mybatis.spring.scan.filter.datasource",
+  @MapperScan(basePackages = "org.mybatis.spring.filter.datasource",
     excludeFilters = {@ComponentScan.Filter(type = FilterType.REGEX,
-      pattern = "org\\.mybatis\\.spring\\.scan\\.filter\\.datasource\\.datasource1\\..*")})
+      pattern = "org\\.mybatis\\.spring\\.filter\\.datasource\\.datasource1\\..*")})
   public static class RegexFilterConfig {
 
   }
 
-  @MapperScan(basePackages = "org.mybatis.spring.scan.filter.datasource",
+  @MapperScan(basePackages = "org.mybatis.spring.filter.datasource",
     excludeFilters = {@ComponentScan.Filter(type = FilterType.ASPECTJ,
       pattern = "*..DataSource1Mapper")})
   public static class AspectJFilterConfig {
