+++ src/test/java/org/mybatis/spring/filter/customfilter/AnnoTypeFilter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.mybatis.spring.scan.filter.customfilter;
+package org.mybatis.spring.filter.customfilter;
 
 import java.lang.annotation.*;
