+++ src/test/java/org/mybatis/spring/filter/datasource/commonsource/AnnoExcludeMapper.java
@@ -13,7 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.mybatis.spring.scan.filter.datasource.datasource2;
+package org.mybatis.spring.filter.datasource.commonsource;
 
-public interface DataSource2Mapper {
+import org.mybatis.spring.filter.customfilter.AnnoTypeFilter;
+
+@AnnoTypeFilter
+public interface AnnoExcludeMapper {
 }
